Add pagination get posts API

+++ backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res
   const post  = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()
   .then(post => {
@@ -81,14 +82,16 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  Post.updateOne({ _id: req.params.id}, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
   .then(result => {
-    res.status(200).json({
-      message: 'Post updated successfully!',
-      posts: result
-    })
+    if(result.nModified > 0) {
+      res.status(200).json({message: 'Post updated successfully!' })
+    } else {
+      res.status(401).json({message: 'Not Authorized!' })
+    }
   })
 })
 
@@ -103,12 +106,13 @@ router.get('/:id', (req, res, next)=> {
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
   .then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Post deleted!'
-    })
+    if(result.n > 0) {
+      res.status(200).json({message: 'Post deleted successfully!' })
+    } else {
+      res.status(401).json({message: 'Not Authorized!' })
+    }
   })
 })
